[IMP] l10n_it_dichiarazione_intento: Translate all the things to english
