+++ l10n_it_declaration_of_intent/i18n/l10n_it_dichiarazione_intento.pot
