+++ l10n_it_declaration_of_intent/models/__init__.py
@@ -1,0 +1,6 @@
+from . import account
+from . import declaration
+from . import company
+from . import account_move
+from . import sale
+from . import product
