+++ l10n_it_declaration_of_intent/readme/USAGE.rst
