+++ l10n_it_declaration_of_intent/security/ir.model.access.csv
@@ -1,0 +1,6 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_declaration_of_intent,declaration_of_intent access,model_l10n_it_declaration_of_intent_declaration,account.group_account_invoice,1,1,1,1
+access_declaration_of_intent_base,declaration_of_intent read access,model_l10n_it_declaration_of_intent_declaration,base.group_user,1,0,0,0
+access_declaration_of_intent_line,declaration_of_intent line access,model_l10n_it_declaration_of_intent_declaration_line,account.group_account_invoice,1,1,1,1
+access_declaration_of_intent_yearly_limit,declaration_of_intent.yearly_limit access,model_l10n_it_declaration_of_intent_yearly_limit,account.group_account_invoice,1,1,1,1
+access_select_declarations,access_select_declarations,model_l10n_it_declaration_of_intent_select_declarations,base.group_user,1,1,1,1
