+++ l10n_it_declaration_of_intent/static/description/index.html
