+++ l10n_it_declaration_of_intent/tests/__init__.py
@@ -1,0 +1,5 @@
+# Copyright 2017 Francesco Apruzzese <f.apruzzese@apuliasoftware.it>
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from . import test_account
+from . import test_declaration_of_intent
