+++ l10n_it_declaration_of_intent/views/account_invoice_view.xml
@@ -1,0 +1,51 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="declaration_of_intent_invoice_form" model="ir.ui.view">
+        <field name="name">Add declaration of intent fields to invoice form</field>
+        <field name="model">account.move</field>
+        <field name="inherit_id" ref="account.view_move_form" />
+        <field name="arch" type="xml">
+            <field name="fiscal_position_id" position="after">
+                <label for="declaration_of_intent_ids" />
+                <div>
+                    <field
+                        name="declaration_of_intent_ids"
+                        widget="many2many_tags"
+                        readonly="1"
+                        class="oe_inline oe_link"
+                    />
+                    <button
+                        type="object"
+                        name="select_manually_declarations"
+                        states="draft"
+                        string="Select Manually"
+                        class="oe_inline oe_link"
+                    />
+                </div>
+            </field>
+            <xpath
+                expr="//field[@name='invoice_line_ids']/tree/field[@name='tax_ids']"
+                position="after"
+            >
+                <field
+                    name="force_declaration_of_intent_id"
+                    domain="[('partner_id', '=', parent.partner_id), ('state', '!=', 'close'), ('date_start', '&lt;=', parent.invoice_date), ('date_end', '&gt;=', parent.invoice_date)]"
+                />
+            </xpath>
+        </field>
+    </record>
+
+    <record id="declaration_of_intent_invoice_line_form" model="ir.ui.view">
+        <field name="name">Add declaration of intent fields to invoice line form</field>
+        <field name="model">account.move.line</field>
+        <field name="inherit_id" ref="account.view_move_line_form" />
+        <field name="arch" type="xml">
+            <field name="tax_ids" position="after">
+                <field
+                    name="force_declaration_of_intent_id"
+                    domain="[('partner_id', '=', parent.partner_id), ('state', '!=', 'close'), ('date_start', '&lt;=', parent.invoice_date), ('date_end', '&gt;=', parent.invoice_date)]"
+                />
+            </field>
+        </field>
+    </record>
+</odoo>
