+++ l10n_it_declaration_of_intent/views/account_view.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="declaration_of_intent_account_position_form" model="ir.ui.view">
+        <field name="name">Add declaration of intent fields</field>
+        <field name="model">account.fiscal.position</field>
+        <field name="inherit_id" ref="account.view_account_position_form" />
+        <field name="arch" type="xml">
+            <field name="auto_apply" position="after">
+                <field name="valid_for_declaration_of_intent" />
+            </field>
+        </field>
+    </record>
+</odoo>
