+++ l10n_it_declaration_of_intent/views/company_view.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="declaration_of_intent_view_company_form" model="ir.ui.view">
+        <field name="name">Add declaration of intent fields</field>
+        <field name="model">res.company</field>
+        <field name="inherit_id" ref="base.view_company_form" />
+        <field name="arch" type="xml">
+            <xpath expr="//page[last()]" position="after">
+                <page string="Declarations of intent">
+                    <field name="declaration_yearly_limit_ids">
+                        <tree editable="top">
+                            <field name="year" />
+                            <field name="limit_amount" />
+                            <field name="used_amount" />
+                        </tree>
+                    </field>
+                </page>
+            </xpath>
+        </field>
+    </record>
+</odoo>
