+++ l10n_it_declaration_of_intent/wizard/__init__.py
@@ -1,0 +1,1 @@
+from . import select_declarations
