+++ l10n_it_declaration_of_intent/wizard/select_declarations_view.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="select_manually_declarations_wizard" model="ir.ui.view">
+        <field name="name">Select Manually Declarations</field>
+        <field name="model">l10n_it_declaration_of_intent.select_declarations</field>
+        <field name="arch" type="xml">
+            <form string="Select Manually Declarations">
+                <group>
+                    <field name="declaration_ids" nolabel="1" />
+                </group>
+                <footer>
+                    <button
+                        name="confirm"
+                        string="Confirm"
+                        type="object"
+                        class="oe_highlight"
+                    />
+                    or
+                    <button string="Cancel" class="oe_link" special="cancel" />
+                </footer>
+            </form>
+        </field>
+    </record>
+
+    <record id="select_manually_declarations_action" model="ir.actions.act_window">
+        <field name="name">Select Manually Declarations</field>
+        <field
+            name="res_model"
+        >l10n_it_declaration_of_intent.select_declarations</field>
+        <field name="view_mode">tree,form</field>
+        <field name="view_id" ref="select_manually_declarations_wizard" />
+        <field name="target">new</field>
+    </record>
+</odoo>
