--- l10n_it_dichiarazione_intento/models/__init__.py
@@ -1,6 +1,0 @@
-from . import account
-from . import dichiarazione_intento
-from . import company
-from . import account_invoice
-from . import sale
-from . import product
