--- l10n_it_dichiarazione_intento/tests/__init__.py
@@ -1,5 +1,0 @@
-# Copyright 2017 Francesco Apruzzese <f.apruzzese@apuliasoftware.it>
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from . import test_account
-from . import test_dichiarazione_intento
