--- l10n_it_dichiarazione_intento/views/account_view.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <record id="dichiarazione_intento_account_position_form" model="ir.ui.view">
-        <field name="name">dichiarazione.intento.account.fiscal.position.form</field>
-        <field name="model">account.fiscal.position</field>
-        <field name="inherit_id" ref="account.view_account_position_form" />
-        <field name="arch" type="xml">
-            <field name="auto_apply" position="after">
-                <field name="valid_for_dichiarazione_intento" />
-            </field>
-        </field>
-    </record>
-</odoo>
