--- l10n_it_dichiarazione_intento/views/company_view.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <record id="dichiarazione_intento_view_company_form" model="ir.ui.view">
-        <field name="name">dichiarazione.intento.res.company.form</field>
-        <field name="model">res.company</field>
-        <field name="inherit_id" ref="base.view_company_form" />
-        <field name="arch" type="xml">
-            <xpath expr="//page[last()]" position="after">
-                <page string="Declarations of intent">
-                    <field name="dichiarazione_yearly_limit_ids">
-                        <tree editable="top">
-                            <field name="year" />
-                            <field name="limit_amount" />
-                            <field name="used_amount" />
-                        </tree>
-                    </field>
-                </page>
-            </xpath>
-        </field>
-    </record>
-</odoo>
