--- l10n_it_dichiarazione_intento/wizard/__init__.py
@@ -1,1 +1,0 @@
-from . import manually_declarations
