+++ setup/l10n_it_declaration_of_intent/odoo/addons/l10n_it_declaration_of_intent
@@ -1,0 +1,1 @@
+../../../../l10n_it_declaration_of_intent
