--- setup/l10n_it_dichiarazione_intento/odoo/addons/l10n_it_dichiarazione_intento
@@ -1,1 +1,0 @@
-../../../../l10n_it_dichiarazione_intento
